+++ llama_cpp/llama_cpp.py
@@ -15,28 +15,32 @@ from ctypes import (
     c_size_t,
 )
 import pathlib
+from typing import List
 
 
 # Load the library
 def _load_shared_library(lib_base_name: str):
-    # Determine the file extension based on the platform
-    if sys.platform.startswith("linux"):
-        lib_ext = ".so"
-    elif sys.platform == "darwin":
-        lib_ext = ".so"
-    elif sys.platform == "win32":
-        lib_ext = ".dll"
-    else:
-        raise RuntimeError("Unsupported platform")
-
     # Construct the paths to the possible shared library names
     _base_path = pathlib.Path(__file__).parent.resolve()
     # Searching for the library in the current directory under the name "libllama" (default name
     # for llamacpp) and "llama" (default name for this repo)
-    _lib_paths = [
-        _base_path / f"lib{lib_base_name}{lib_ext}",
-        _base_path / f"{lib_base_name}{lib_ext}",
-    ]
+    _lib_paths: List[pathlib.Path] = []
+    # Determine the file extension based on the platform
+    if sys.platform.startswith("linux"):
+        _lib_paths += [
+            _base_path / f"lib{lib_base_name}.so",
+        ]
+    elif sys.platform == "darwin":
+        _lib_paths += [
+            _base_path / f"lib{lib_base_name}.so",
+            _base_path / f"lib{lib_base_name}.dylib",
+        ]
+    elif sys.platform == "win32":
+        _lib_paths += [
+            _base_path / f"{lib_base_name}.dll",
+        ]
+    else:
+        raise RuntimeError("Unsupported platform")
 
     if "LLAMA_CPP_LIB" in os.environ:
         lib_base_name = os.environ["LLAMA_CPP_LIB"]
@@ -78,6 +82,10 @@ c_uint8_p = POINTER(c_uint8)
 c_size_t_p = POINTER(c_size_t)
 
 # llama.h bindings
+
+GGML_USE_CUBLAS = hasattr(_lib, "ggml_init_cublas")
+GGML_CUDA_MAX_DEVICES = ctypes.c_int(16)
+LLAMA_MAX_DEVICES = GGML_CUDA_MAX_DEVICES if GGML_USE_CUBLAS else ctypes.c_int(1)
 
 # #define LLAMA_FILE_MAGIC_GGJT        0x67676a74u // 'ggjt'
 LLAMA_FILE_MAGIC_GGJT = ctypes.c_uint(0x67676A74)
@@ -142,9 +150,12 @@ llama_progress_callback = ctypes.CFUNCTYPE(None, c_float, c_void_p)
 
 
 # struct llama_context_params {
-#     int n_ctx;        // text context
-#     int n_gpu_layers; // number of layers to store in VRAM
-#     int seed;         // RNG seed, -1 for random
+#     int n_ctx;                             // text context
+#     int n_batch;                           // prompt processing batch size
+#     int n_gpu_layers;                      // number of layers to store in VRAM
+#     int main_gpu;                          // the GPU that is used for scratch and small tensors
+#     float tensor_split[LLAMA_MAX_DEVICES]; // how to split layers across multiple GPUs
+#     int seed;                              // RNG seed, -1 for random
 
 #     bool f16_kv;     // use fp16 for KV cache
 #     bool logits_all; // the llama_eval() call computes all logits, not just the last one
@@ -162,7 +173,10 @@ llama_progress_callback = ctypes.CFUNCTYPE(None, c_float, c_void_p)
 class llama_context_params(Structure):
     _fields_ = [
         ("n_ctx", c_int),
+        ("n_batch", c_int),
         ("n_gpu_layers", c_int),
+        ("main_gpu", c_int),
+        ("tensor_split", c_float * LLAMA_MAX_DEVICES.value),
         ("seed", c_int),
         ("f16_kv", c_bool),
         (