+++ CHANGELOG.md
@@ -7,24 +7,48 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Added: k-quants support
+## [0.1.60]
+
+#### Addeed
+
+- (build-system) Migrate from scikit-build to scikit-build-core
+
+### Fixed
+
+- Truncate max_tokens in create_completion so requested tokens doesn't exceed context size.
+- Temporarily disable cache for completion requests
+
+## [v0.1.59]
+
+### Added
+
+- (llama.cpp) k-quants support
+- (server) mirostat sampling parameters to server
+
+### Fixed
+
+- Support both `.so` and `.dylib` for `libllama` on MacOS
 
 ## [v0.1.58]
 
-- Added: Metal Silicon support
+### Added
+
+- (llama.cpp) Metal Silicon support
 
 ## [v0.1.57]
 
-- Added: OpenLlama 3B support
+### Added
+
+- (llama.cpp) OpenLlama 3B support
 
 ## [v0.1.56]
 
 ### Added
 
-- Added first version of the changelog
-- Server: Use async routes
-- Use numpy for internal buffers to reduce memory usage and improve performance.
+- (misc) Added first version of the changelog
+- (server) Use async routes
+- (python-api) Use numpy for internal buffers to reduce memory usage and improve performance.
 
 ### Fixed
 
-- Performance bug in stop sequence check slowing down streaming.
+- (python-api) Performance bug in stop sequence check slowing down streaming.